tests: move basic functionality tests to "test_sanity"
