+++ cli/tests/integrations/test_dcos.py
@@ -1,4 +1,4 @@
-from .common import assert_command, exec_command
+from .helpers.common import assert_command, exec_command
 
 
 def test_default():
