+++ cli/tests/integrations/test_service.py
@@ -9,11 +9,12 @@ import pytest
 import dcos.util as util
 from dcos.util import create_schema
 
-from .common import (assert_command, assert_lines, delete_zk_node,
-                     delete_zk_nodes, exec_command, get_services,
-                     package, package_install, remove_app, service_shutdown,
-                     setup_universe_server, ssh_output,
-                     teardown_universe_server, wait_for_service)
+from .helpers.common import (assert_command, assert_lines, delete_zk_node,
+                             delete_zk_nodes, exec_command, ssh_output)
+from .helpers.marathon import remove_app
+from .helpers.package import (package, package_install, setup_universe_server,
+                              teardown_universe_server)
+from .helpers.service import get_services, service_shutdown, wait_for_service
 from ..fixtures.service import framework_fixture
